remove unnecessary cast

+++ src/main.cpp
@@ -16,10 +16,8 @@ int main() {
     for (int row = 0; row < height; ++row) {
         std::clog << "Scanlines remaining: " << (height - row) << std::endl;
         for (int col = 0; col < width; ++col) {
-            const double red =
-                static_cast<double>(row) / static_cast<double>(height - 1);
-            const double blue =
-                static_cast<double>(col) / static_cast<double>(width - 1);
+            const double red = static_cast<double>(row) / (height - 1);
+            const double blue = static_cast<double>(col) / (width - 1);
             Color color(red, 0.0, blue);
 
             writeColor(std::cout, color);
